requirements.txt
streamlit==1.28.0
pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.8.2
tensorflow==2.15.0
flask==3.0.0 